requirements.txt
together
aiohttp
aiosqlite==0.20.0
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.8.2
web3==6.15.1

hive-agent @ git+https://github.com/hivenetwork-ai/hive-agent-py@main
